(feature/SupportMinGw64GccBuild) Supported masp

+++ Programs/Sources/Libraries/masp/Basics/Iir/main.cpp
@@ -1,3 +1,27 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2015-2016 Ryosuke Kanata
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
 #include <stdio.h>
 #include <math.h>
 #include <stdint.h>
@@ -84,7 +108,7 @@ static void test_filtering(void)
     printf("BEF: b0=%f,b1=%f,b2=%f,a1=%f,a2=%f\n",
         coef.b[0], coef.b[1], coef.b[2], coef.a[0], coef.a[1]);
 
-    for (uint i = 0; i < sweep.GetLength(); ++i)
+    for (size_t i = 0; i < sweep.GetLength(); ++i)
     {
         double audioIn = static_cast<double>(sweep[i]);
         double v = lpf.ApplyFilter(audioIn); sweep_filtered_lpf[i] = static_cast<int16_t>(v);
@@ -129,7 +153,7 @@ int main(void)
     if(0)
     {
         int freqs[] = { Melody::C, Melody::E, Melody::GA, 0 };
-        for (uint i = 0; i < sizeof(freqs)/sizeof(freqs[0]) - 1; ++i)
+        for (size_t i = 0; i < sizeof(freqs)/sizeof(freqs[0]) - 1; ++i)
         {
             int freq = static_cast<int>(Melody::GetRate(freqs[i]) + 0.5);
             printf("%i\n", freq);
