+++ Programs/Sources/Libraries/mtbx/WaveGen/WaveGen.cpp
@@ -225,7 +225,7 @@ double WaveGen::GenerateOne(double p)
 
 void WaveGen::GenerateWaveform(mcon::Vector<double>& buffer, double amp)
 {
-    for (uint i = 0; i < buffer.GetLength(); ++i, ++(*this))
+    for (size_t i = 0; i < buffer.GetLength(); ++i, ++(*this))
     {
         buffer[i] = amp * GetValue();
     }
